Android port (remove dbus)
Since dbus is no longer used in Android by any component, it may be
removed soon.
The android_main.cpp does not use any dbus interface. Still kept a
version supporting dbus, which will be modified to use binder.
In this process also removed some warnings for size_t variation in
bionic and glib.

+++ src/thd_trip_point.cpp
@@ -122,7 +122,8 @@ bool cthd_trip_point::thd_trip_point_check(int id, unsigned int read_temp,
 		return true;
 
 	int i, ret;
-	thd_log_debug("cdev size for this trippoint %lu\n", cdevs.size());
+	thd_log_debug("cdev size for this trippoint %lu\n",
+			(unsigned long) cdevs.size());
 	if (on > 0) {
 		for (unsigned i = 0; i < cdevs.size(); ++i) {
 			cthd_cdev *cdev = cdevs[i].cdev;
